# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: COSN API Gateway
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
  - url: https://cosn-api-gateway.herokuapp.com
    description: Heroku server url
paths:
  /runVisualization:
    post:
      tags:
        - Visualization Service
      operationId: runVisualization
      parameters:
        - name: AuthToken
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisualizationRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /register:
    post:
      tags:
        - Authentication
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /login:
    post:
      tags:
        - Authentication
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /getUserFromToken:
    post:
      tags:
        - Authentication
      operationId: getUserFromToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserFromTokenRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /executeOrder:
    post:
      tags:
        - MAS Service
      operationId: executeOrder
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /changeUserRole:
    post:
      tags:
        - Roles Managment
      operationId: changeUserRole
      parameters:
        - name: AuthToken
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleChangeRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /roles:
    get:
      tags:
        - Roles Managment
      operationId: getAllRoles
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /role/{roleId}:
    get:
      tags:
        - Roles Managment
      operationId: getRoleById
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /getSensorValues:
    get:
      tags:
        - Real System Service
      operationId: getSensorValues
      parameters:
        - name: AuthToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /getDomainModel:
    get:
      tags:
        - Domain Model Service
      operationId: getDomainModel
      parameters:
        - name: AuthToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /getAllOrders:
    get:
      tags:
        - MAS Service
      operationId: getAllOrders
      parameters:
        - name: AuthToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /getAlgorithms:
    get:
      tags:
        - Algorithms Service
      operationId: getAlgorithms
      parameters:
        - name: AuthToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "226":
          description: IM Used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    SimpleResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    VisualizationRequest:
      type: object
      properties:
        tempRequest:
          type: string
    UserRegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        roleId:
          type: integer
          format: int64
    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    UserFromTokenRequest:
      type: object
      properties:
        token:
          type: string
    OrderRequest:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    UserRoleChangeRequest:
      type: object
      properties:
        email:
          type: string
        roleId:
          type: integer
          format: int64